Add Record Serializer to show records list
This adds the JSONSerializer to the gem for an easy way to meet the
JSONAPI standard[1]. That means we can use a simple serialiser for the
records JSON response, defining the fields returned.

Overriding the Condition and Artist to_s methods also gives a nicely
structured record object with only the things we need in the response.

1. https://jsonapi.org/