Add front end to add Records
This allows Records to be added to the collection. The Artist will be
duplicated, but we will add a dropdown to select existing artists to
solve that in an easy way.

Using nested attributes, we can create a new artists if they don't exist
from this form, saving need to add any artists separately (only needing
to edit Artists which could be done in the edit record form or its own
controller).