Add edit record feature
This commit adds the ability to edit a record, adding controller actions
and front end components. The Form was moved to a component as its now
being reused for the add and edit features.

The tests have dropped off a little bit as I'm running out of time and
I'd rather get a few more features working. Earlier work hopefully shows
that test driven development is the default rather than just adding
features...